+++ jackson2/src/main/java/io/github/dfa1/typesafe/jackson2/Jackson2Codec.java
@@ -3,6 +3,7 @@ package io.github.dfa1.typesafe.jackson2;
 import io.github.dfa1.typesafe.core.Answer;
 import io.github.dfa1.typesafe.core.Model;
 import io.github.dfa1.typesafe.core.Question;
+import io.github.dfa1.typesafe.core.RequestId;
 import io.github.dfa1.typesafe.core.State;
 import io.github.dfa1.typesafe.json.JsonCodec;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
@@ -28,7 +29,9 @@ public final class Jackson2Codec implements JsonCodec {
             .registerModule(new SimpleModule()
                     .addSerializer(State.class, new StateSerializer())
                     .addSerializer(Model.class, new ModelSerializer())
-                    .addDeserializer(Model.class, new ModelDeserializer()));
+                    .addDeserializer(Model.class, new ModelDeserializer())
+                    .addSerializer(RequestId.class, new RequestIdSerializer())
+                    .addDeserializer(RequestId.class, new RequestIdDeserializer()));
 
     @Override
     public String writeValueAsString(Object value) {